requirements.txt
streamlit
pandas
sqlalchemy
geopandas
plotly.express
folium
streamlit_folium
matplotlib
psycopg2-binary
Pillow